TIP-674: move product values from model folder
