+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -26,11 +26,3 @@ parameters:
     pim_catalog.entity.metric.class:                       Pim\Component\Catalog\Model\Metric
     pim_catalog.entity.product_price.class:                Pim\Component\Catalog\Model\ProductPrice
     pim_catalog.entity.product_template.class:             Pim\Bundle\CatalogBundle\Entity\ProductTemplate
-
-    # move that ?
-    pim_catalog.entity.product_value.class: Pim\Component\Catalog\Model\ProductValue\ProductValue
-    pim_catalog.entity.product_value.media.class: Pim\Component\Catalog\Model\ProductValue\MediaProductValue
-    pim_catalog.entity.product_value.metric.class: Pim\Component\Catalog\Model\ProductValue\MetricProductValue
-    pim_catalog.entity.product_value.option.class: Pim\Component\Catalog\Model\ProductValue\OptionProductValue
-    pim_catalog.entity.product_value.options.class: Pim\Component\Catalog\Model\ProductValue\OptionsProductValue
-    pim_catalog.entity.product_value.price_collection.class: Pim\Component\Catalog\Model\ProductValue\PriceCollectionProductValue
