+++ src/Pim/Bundle/EnrichBundle/spec/Normalizer/ProductViolationNormalizerSpec.php
@@ -5,8 +5,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Normalizer;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ProductInterface;
-use Pim\Component\Catalog\Model\ProductValue;
-use Pim\Component\Catalog\Model\ProductValueCollection;
+use Pim\Component\Catalog\ProductValue\ProductValue;
 use Pim\Component\Catalog\Model\ProductValueInterface;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
 use Symfony\Component\Validator\ConstraintViolationInterface;
