+++ src/Pim/Component/Catalog/spec/Factory/ProductValue/MediaProductValueFactorySpec.php
@@ -9,7 +9,7 @@ use Akeneo\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Factory\ProductValue\MediaProductValueFactory;
 use Pim\Component\Catalog\Model\AttributeInterface;
-use Pim\Component\Catalog\Model\ProductValue;
+use Pim\Component\Catalog\ProductValue\ProductValue;
 use Prophecy\Argument;
 
 class MediaProductValueFactorySpec extends ObjectBehavior
